refactor(game): move TextureDynamic to separate file
