+++ src/game/graphics/texture_extra.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <future>
 #include <map>
 #include <shared_mutex>
 #include <utility>
@@ -79,24 +78,4 @@ public:
     void stopUpdate();
 
     void setVideoSpeed();
-};
-
-///////////////////////////////////////////////////////////////////////////////
-
-class TextureDynamic final : public Texture
-{
-private:
-    std::future<Image> _loadImage;
-    std::unique_ptr<Texture> _dynTexture;
-    Path _loaded_path;
-
-public:
-    TextureDynamic();
-    ~TextureDynamic() override = default;
-
-    void setPath(const Path& path); // Asynchronously load the image from disk.
-    void applyImageIfNeeded();      // Synchronously upload the image to GPU.
-
-    void draw(const Rect& srcRect, RectF dstRect, const Color c, const BlendMode blend, const bool filter,
-              const double angleInDegrees, const Point* center) const override;
 };
